Custom stop zone overhaul done
