+++ overlay.html
@@ -106,6 +106,24 @@
             border-color: rgba(34, 197, 94, 0.95);
             border-style: dashed;
         }
+        .zone--labeled {
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            overflow: hidden;
+        }
+        .zone-label {
+            font-family: "RobotoFlex", sans-serif;
+            font-size: 10rem;
+            font-weight: 800;
+            line-height: 1;
+            user-select: none;
+        }
+        .zone-custom .zone-label { color: rgba(255, 196, 0, 0.55); }
+        .zone-pause .zone-label  { color: rgba(0, 180, 255, 0.55); }
+        .zone-start .zone-label  { color: rgba(34, 197, 94, 0.55); }
+        .zone-disabled .zone-label { color: rgba(128, 128, 128, 0.55); }
+
         @keyframes fadeOut {
             0% { opacity: 1; }
             70% { opacity: 1; }
@@ -114,6 +132,7 @@
         .fade-out {
             animation: fadeOut 5s ease forwards;
         }
+
     </style>
 </head>
 <body>
@@ -140,6 +159,8 @@
         const stopZonePrompt = document.getElementById('stop-zone-prompt');
         let pickModeActive = false;
         let stopZonePickModeActive = false;
+        let zoneFadeAnim = null;
+        let hasFadedIn = false;
 
         function setupTauriListener() {
             if (!window.__TAURI__ || !window.__TAURI__.event) {
@@ -154,7 +175,10 @@
 
             window.__TAURI__.event.listen('clear-zones', () => {
                 zoneLayer.innerHTML = '';
-                zoneLayer.classList.remove('fade-out');
+                if (zoneFadeAnim) {
+                    zoneFadeAnim.cancel();
+                    zoneFadeAnim = null;
+                }
             });
 
             window.__TAURI__.event.listen('click-points-data', (event) => {
@@ -186,13 +210,23 @@
 
         function renderZones(data) {
             zoneLayer.innerHTML = '';
-            zoneLayer.classList.remove('fade-out');
-            void zoneLayer.offsetWidth;
-            if (!stopZonePickModeActive) {
-                zoneLayer.classList.add('fade-out');
+            if (zoneFadeAnim) {
+                if (zoneFadeAnim.playState === 'finished') {
+                    hasFadedIn = false;
+                }
+                zoneFadeAnim.cancel();
+                zoneFadeAnim = null;
             }
 
             if (!data) return;
+
+            if (!stopZonePickModeActive) {
+                zoneFadeAnim = zoneLayer.animate([
+                    { opacity: 1 },
+                    { opacity: 1, offset: 0.5 },
+                    { opacity: 0 }
+                ], { duration: 3000, easing: 'ease', fill: 'forwards' });
+            }
 
             const sw = data.screenWidth;
             const sh = data.screenHeight;
@@ -205,16 +239,26 @@
             }
 
             if (Array.isArray(data.stopZones)) {
-                for (const zone of data.stopZones) {
+                const disabled = data.stopZonesEnabled === false;
+                data.stopZones.forEach((zone, index) => {
                     let cls = 'zone-custom';
                     if (zone.action === 'pause') cls = 'zone-pause';
                     else if (zone.action === 'start') cls = 'zone-start';
-                    addZone(zone.x, zone.y, zone.width, zone.height, cls);
-                }
+                    if (disabled) cls += ' zone-disabled';
+                    addZone(zone.x, zone.y, zone.width, zone.height, cls, String(index + 1));
+                });
             }
 
             for (const monitor of monitors) {
                 renderMonitorZones(monitor, data);
+            }
+
+            if (!hasFadedIn) {
+                hasFadedIn = true;
+                zoneLayer.animate([
+                    { opacity: 0 },
+                    { opacity: 1 }
+                ], { duration: 250, easing: 'ease' });
             }
         }
 
@@ -281,7 +325,7 @@
             }
         }
 
-        function addZone(x, y, w, h, cls) {
+        function addZone(x, y, w, h, cls, label) {
             const scale = window.devicePixelRatio || 1;
             x = x / scale;
             y = y / scale;
@@ -295,6 +339,16 @@
             div.style.top = y + 'px';
             div.style.width = w + 'px';
             div.style.height = h + 'px';
+            if (label !== undefined) {
+                div.classList.add('zone--labeled');
+                const span = document.createElement('span');
+                span.className = 'zone-label';
+                span.textContent = label;
+                const maxPx = 160;
+                const fitPx = Math.min(w, h) * 0.65 / (1 + Math.max(0, label.length - 1) * 0.2);
+                span.style.fontSize = Math.min(fitPx, maxPx) + 'px';
+                div.appendChild(span);
+            }
             zoneLayer.appendChild(div);
             return div;
         }
@@ -376,7 +430,10 @@
             stopZonePickModeActive = active;
             stopZonePrompt.classList.toggle('active', active);
             if (active) {
-                zoneLayer.classList.remove('fade-out');
+                if (zoneFadeAnim) {
+                    zoneFadeAnim.cancel();
+                    zoneFadeAnim = null;
+                }
             } else {
                 clearStopZonePreview();
             }
