+++ src/App.tsx
@@ -72,10 +72,10 @@ type DropdownOverflowDetail = {
 function getPanelSize(tab: Tab, hasUpdate: boolean) {
   const extra = hasUpdate ? 30 : 0;
   if (tab === "simple") {
-    return { width: 716, height: 175 + extra };
+    return { width: 750, height: 175 + extra };
   }
   if (tab === "settings") return { width: 700, height: 720 + extra };
-  if (tab === "zones") return { width: 750, height: 720 + extra };
+  if (tab === "zones") return { width: 750, height: 700 + extra };
   if (tab === "click-points") return { width: 550, height: 600 + extra };
   return { width: 900, height: 469 + extra };
 }
