+++ src/components/panels/zones/ZonesPanel.css
@@ -4,7 +4,7 @@
 
 .adv-zones-row {
   display: flex;
-  gap: 0.625rem;
+  gap: 0.5rem;
   flex: none;
   align-items: stretch;
   min-height: 0;
@@ -14,7 +14,7 @@
   flex: 1;
   display: flex;
   flex-direction: column;
-  gap: 0.625rem;
+  gap: 0.5rem;
   min-width: 0;
 }
 
@@ -35,7 +35,17 @@
 }
 
 .adv-panel--zones {
+  flex: 1;
   overflow: hidden;
+}
+
+.adv-panel--zones .adv-sectioncontainer {
+  box-shadow: none;
+  margin-bottom: 0.5rem;
+}
+
+.adv-panel--zones > .adv-sectioncontainer:last-child {
+  margin-bottom: 0;
 }
 
 .adv-corner-grid {
@@ -197,16 +207,17 @@
 
 .adv-zone-row {
   display: grid;
-  grid-template-columns: 1.5rem 3.25rem 3.25rem 3.25rem 3.25rem 5rem auto 1.5rem;
-  gap: 0.3125rem;
+  grid-template-columns: 1.5rem 4rem 4rem 4rem 4rem 1fr 5rem 8rem 2rem;
+  gap: 0.25rem;
   align-items: center;
   border: 0.0625rem solid var(--border-subtle);
   border-radius: var(--radius-sm);
-  padding: 0.3125rem 0.375rem;
+  padding: 0.25rem 0.5rem;
   background: var(--bg-elevated);
   width: 100%;
   box-sizing: border-box;
   flex: 0 0 auto;
+  min-height: 2.5rem;
 }
 
 .adv-zone-index {
@@ -217,24 +228,30 @@
   text-align: center;
   width: 1.5rem;
   opacity: 1;
+  grid-column: 1;
 }
 
 .adv-zone-coord {
   gap: 0.125rem;
-  padding: 0 0.3rem;
+  padding: 0 0.25rem;
   min-width: 0;
 }
 
 .adv-zone-coord .adv-number-sm {
-  max-width: 4ch;
-  padding: 0;
+  max-width: 7ch;
+  width: 7ch;
   text-overflow: clip;
 }
 
+.adv-zone-row .adv-zone-draw-btn {
+  grid-column: 8;
+  width: 100%;
+}
+
 .adv-zone-draw-btn {
-  height: 1.5rem;
-  padding: 0 0.375rem;
-  font-size: 0.7rem;
+  height: 2rem;
+  padding: 0 0.5rem;
+  font-size: 0.75rem;
   line-height: 1;
   white-space: nowrap;
   border: 0.0625rem solid var(--border-subtle);
@@ -256,17 +273,60 @@
 }
 
 .adv-zone-row .adv-numbox-sm {
-  height: 1.5rem;
-  min-height: 1.5rem;
+  height: 2rem;
+  min-height: 2rem;
 }
 
 .adv-zone-row .adv-number-sm {
-  height: 1.375rem;
-  line-height: 1.375rem;
+  height: 2rem;
+  line-height: 2rem;
+  width: 7ch !important;
+  text-align: right;
+  padding-right: 0.6rem;
 }
 
 .adv-zone-row .adv-dropdown {
+  grid-column: 7;
+  justify-self: end;
   min-width: 0;
+  width: 100%;
+}
+
+.adv-zone-row .adv-dropdown-trigger {
+  height: 2rem;
+  min-height: 2rem;
+  border: 0.0625rem solid var(--border-subtle);
+  border-radius: var(--radius-sm);
+  background: var(--bg-elevated);
+  box-sizing: border-box;
+  padding: 0 0.5rem;
+}
+
+.adv-zone-row .adv-dropdown-trigger:hover {
+  border-color: var(--accent-ring);
+}
+
+.adv-zone-row .adv-click-points-delete {
+  grid-column: 9;
+  justify-self: end;
+  width: 2rem;
+  height: 2rem;
+  background: transparent;
+  border-radius: var(--radius-sm);
+  border: 0.0625rem solid var(--border-subtle);
+  color: var(--text-muted);
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  transition:
+    color 0.15s ease,
+    border-color 0.15s ease;
+}
+
+.adv-zone-row .adv-click-points-delete:hover {
+  color: #e24d4d;
+  border-color: #e24d4d;
 }
 
 .adv-zones-empty {
@@ -275,8 +335,108 @@
   align-items: center;
   justify-content: center;
   gap: 0.375rem;
+  flex: 1;
+  min-height: 0;
   padding: 1.5rem 1rem;
   text-align: center;
+}
+
+.adv-stop-zones-card {
+  display: flex;
+  flex-direction: column;
+  min-height: 0;
+  overflow: hidden;
+}
+
+.adv-stop-zones-card .adv-disabled-container {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  min-height: 0;
+}
+
+.adv-stop-zones-card .adv-disabled-container > :first-child {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  min-height: 0;
+}
+
+.adv-stop-zones-card .adv-disabled-content {
+  flex: 1;
+  min-height: 0;
+}
+
+.adv-stop-zone-body {
+  display: flex;
+  flex-direction: column;
+  gap: 0.625rem;
+  align-items: stretch;
+  min-height: 0;
+  height: 100%;
+}
+
+.adv-stop-zone-controls {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  min-width: 0;
+  flex: 1;
+  width: 100%;
+  min-height: 0;
+}
+
+.adv-stop-zone-list-shell {
+  position: relative;
+  width: 100%;
+  flex: 1 1 auto;
+  height: 0;
+  min-height: 0;
+}
+
+.adv-stop-zone-list {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  z-index: 0;
+  width: 100%;
+  overflow-y: hidden;
+  min-height: 0;
+  height: 100%;
+  padding: 0 2px 0 0;
+  box-sizing: border-box;
+  scrollbar-gutter: stable;
+  animation: adv-cplist-enable 0.01s 350ms forwards;
+}
+
+.adv-stop-zone-list::-webkit-scrollbar {
+  width: 0.375rem;
+  background: var(--bg-elevated);
+  border-radius: var(--radius-md);
+}
+
+.adv-stop-zone-list::-webkit-scrollbar-thumb {
+  background: var(--text-dim);
+  border-radius: var(--radius-sm);
+}
+
+.adv-stop-zone-list::-webkit-scrollbar-thumb:hover {
+  background: #777;
+}
+
+.adv-stop-zone-list > * + * {
+  margin-top: 6px;
+}
+
+.adv-stop-zone-list-fade {
+  position: absolute;
+  z-index: 2;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  height: 1.75rem;
+  pointer-events: none;
+  background: linear-gradient(to bottom, transparent 0%, var(--bg-base) 100%);
 }
 
 .adv-zones-empty-icon {
